==> 4 _Cleaning YouTube metadata.ipynb <==
The purpose of the following code was to prepare the videos' metadata, which was generated in notebook 2 (Scraping YouTube metadata), for later analyses and calculations. Parts of this code were generated with the help of ChatGPT and altered for the specific needs of this study.

import pandas as pd

BILD = pd.read_csv("BILD_video_data_cleaned.csv")
DW = pd.read_csv("DW_video_data_cleaned.csv")

print(BILD.head())
print(DW.head())


BILD["year"] = BILD["published_at"].str[:4].astype(int)
DW["year"] = DW["published_at"].str[:4].astype(int)

BILD = BILD[BILD["year"].between(2020, 2023)]
DW = DW[DW["year"].between(2020, 2023)]

BILD = BILD.drop(columns=["year"])
DW = DW.drop(columns=["year"])


#extract hashtags

import re

def extract_hashtags(text):
    if isinstance(text, str):
        return re.findall(r"#\w+", text)
    return []

BILD["hashtags"] = BILD["description"].apply(extract_hashtags)
DW["hashtags"] = DW["description"].apply(extract_hashtags)

print(BILD[["description", "hashtags"]].head())
print(DW[["description", "hashtags"]].head())


BILD_description_end_phrases = [
    "BILD IM TV",
    "Abonniert jetzt unseren Kanal und aktiviert die Glocke!",
    "JETZT Sendersuchlauf starten",
    "► Die ganze Sendung hier:"
]

DW_description_end_phrases = [
    "DW Deutsch Abonnieren"
]

def make_cut_function(phrases):
    pattern = r"|".join(map(re.escape, phrases))
    escaped_phrases = [re.escape(p) for p in phrases if p != "------------"]
    dash_regex = r"-{12,}"

    pattern_parts = escaped_phrases + [dash_regex]
    pattern = r"|".join(pattern_parts)

    def cut_description(text):
        if not isinstance(text, str):
            return text
        return re.split(pattern, text)[0].strip()

    return cut_description

cut_BILD = make_cut_function(BILD_description_end_phrases)
cut_DW = make_cut_function(DW_description_end_phrases)

BILD["description"] = BILD["description"].apply(cut_BILD)
DW["description"] = DW["description"].apply(cut_DW)

# concatenating title and description for future topic classification
BILD["text"] = BILD["title"].astype(str) + ": " + BILD["description"].astype(str)
DW["text"] = DW["title"].astype(str) + ": " + DW["description"].astype(str)

BILD["description"] = BILD["description"].apply(cut_BILD)
DW["description"] = DW["description"].apply(cut_DW)

BILD.to_csv("BILD_ready_for_classification.csv", index=False)
DW.to_csv("DW_ready_for_classification.csv", index=False)

from google.colab import files

files.download("BILD_ready_for_classification.csv")
files.download("DW_ready_for_classification.csv")
