==> 5 _Cleaning the YouTube comments.ipynb <==
The purpose of the following code was to concatenate BILD's many comment csv files that were fetched in notebook 3_1 (Scraping the YouTube comments) into a single dataset. Additionally, some alterations were made for later evaluations. Parts of this code were generated with the help of ChatGPT and altered for the specific needs of this study.

import pandas as pd

# list of BILD comment files
BILD_files = [f"comments_BILD_{i}.csv" for i in range(1, 12)]

# read + concatenate the files
BILD_comments = pd.concat(
    (pd.read_csv(file, encoding="utf-8") for file in BILD_files),
    ignore_index=True
)

# drop unneccessary columns
BILD_comments = BILD_comments.drop(columns=["author", "updated_at"], errors="ignore")

# make commentors anonymous
BILD_comments["text"] = (
    BILD_comments["text"]
    .astype(str)
    .str.replace(r"@\S+", "", regex=True)
    .str.replace(r"\s+", " ", regex=True)
    .str.strip()
)

BILD_comments


!pip install emoji

import emoji
import pandas as pd

# function to convert the emojis in the comments into text for later sentiment analysis
def demojize_german(text):
    if not isinstance(text, str):
        return text

    return emoji.demojize(text, language="de")

BILD_comments["text"] = BILD_comments["text"].apply(demojize_german)


BILD_comments.to_csv("BILD_comments_full.csv", index=False)

DW_comments   = pd.read_csv("comments_DW.csv", encoding="utf-8")

# drop unneccessary columns
DW_comments = DW_comments.drop(columns=["author", "updated_at"], errors="ignore")

# make commentors anonymous
DW_comments["text"] = (
    DW_comments["text"]
    .astype(str)
    .str.replace(r"@\S+", "", regex=True)
    .str.replace(r"\s+", " ", regex=True)
    .str.strip()
)

DW_comments

DW_comments["text"] = DW_comments["text"].apply(demojize_german)

DW_comments.to_csv("DW_comments_full.csv", index=False)
