==> 9 _Sentiment analysis videos.ipynb <==
The purpose of the following code was to generate sentiment values for each video based on the tables from notebook 8 (Classifying topics of YouTube videos).
Parts of this code were generated with the help of ChatGPT and altered for the specific needs of this study.

import pandas as pd
#sentiment classification for the YouTube videos' title and description

BILD = pd.read_csv("BILD_with_topics.csv", encoding="utf-8")
DW   = pd.read_csv("DW_with_topics.csv", encoding="utf-8")

BILD = BILD.drop_duplicates().reset_index(drop=True)
DW   = DW.drop_duplicates().reset_index(drop=True)

print("BILD:", BILD.shape)
print("DW:", DW.shape)


!pip install transformers torch -q

import pandas as pd
import numpy as np
from transformers import AutoTokenizer, AutoModelForSequenceClassification, pipeline

model_name = "cardiffnlp/twitter-xlm-roberta-base-sentiment"

tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModelForSequenceClassification.from_pretrained(model_name)

sentiment_analyzer = pipeline(
    "sentiment-analysis",
    model=model,
    tokenizer=tokenizer,
    truncation=True,
    max_length=256
)

def preprocess(text):
    return str(text)


def analyze_sentiment(text):
    r = sentiment_analyzer(preprocess(text))[0]
    return r["label"], r["score"]

BILD[["sentiment_label", "sentiment_score"]] = (
    BILD["text"].apply(lambda x: pd.Series(analyze_sentiment(x)))
)

DW[["sentiment_label", "sentiment_score"]] = (
    DW["text"].apply(lambda x: pd.Series(analyze_sentiment(x)))
)


print(BILD[["text", "sentiment_label", "sentiment_score"]].head())
print(DW[["text", "sentiment_label", "sentiment_score"]].head())


from google.colab import files
BILD.to_csv("BILD_final.csv", index=False)
DW.to_csv("DW_final.csv", index=False)

files.download("BILD_final.csv")
files.download("DW_final.csv")
