==> 12 _Concatenating BILD comments.ipynb <==
The purpose of the following code was to concatenate all BILD comment tables generated in notebooks 10_1 and 10_2 into a single CSV file.

import pandas as pd
from pathlib import Path

base_path = Path(r"D:\Users\mague\Desktop\Uni\Bachelor Arbeit\Code\Datasets\Scraped Comments\Comments with sentiment")

csv_files = base_path.rglob("*.csv")

# concat all comment csv files
BILD_comments = pd.concat(
    (pd.read_csv(file, encoding="utf-8") for file in csv_files),
    ignore_index=True
)

BILD_comments = (
    BILD_comments
    .drop_duplicates()
    .reset_index(drop=True)
)


print(BILD_comments.head())
print(BILD_comments.shape)


BILD_comments.to_csv(r"D:\Users\mague\Desktop\Uni\Bachelor Arbeit\Code\Datasets\Scraped Comments\BILD_comments_final.csv", 
          index=False, encoding="utf-8")
